requirements.txt
Flask
google-cloud-aiplatform
gunicorn
python-dotenv
google-generativeai
markdown
requests
tavily-python
python-dotenv